pyproject.toml
# NOTE: you have to use single-quoted strings in TOML for regular expressions.
# It's the equivalent of r-strings in Python.  Multiline strings are treated as
# verbose regular expressions by Black.  Use [ ] to denote a significant space
# character.

[tool.black]
line-length = 88
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
use_parentheses=true
multi_line_output=3
include_trailing_comma=true
line_length=79
known_first_party=['task_processor']
known_third_party=['django', 'rest_framework', 'saml2', 'drf_yasg2', 'pytest']
skip = ['migrations', 'flagsmith', '.venv']

[tool.poetry]
name = "flagsmith_task_processor"
version = "1.0.0"
description = "Task Processor plugin for Flagsmith application."
authors = ["Flagsmith <support@flagsmith.com>"]
readme = "readme.md"
include = [{ path = "migrations/sql/*", format = ["sdist", "wheel"] }]
packages = [{ include = "task_processor"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2.23"
django-health-check = "~3.14.3"
backoff = "~2.2.1"
simplejson = "~3.19.1"
djangorestframework = "~3.12.1"
drf-yasg = "~1.21.6"
dj-database-url = "~0.5.0"
environs = "~9.2.0"
psycopg2-binary = "~2.9.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "~3.0.4"
flake8 = "~6.0.0"
pytest-mock = "~3.10.0"
pylint = "~2.16.2"
pep8 = "~1.7.1"
autopep8 = "~2.0.1"
pytest = "~7.2.1"
black = "~23.7.0"
pytest-django = "^4.6.0"
django-capture-on-commit-callbacks = "^1.11.0"
pytest-freezegun = "~0.4.2"
isort = "~5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
